+++ debug/test.php
@@ -20,17 +20,77 @@ include_once '../src/searchad/search/GeoRequest.php';
 //update
 $id = 10985361;
 $grid = 10985381;
-$api = new \searchad\campaign\CampaignRequest();
+
+$rep = new \searchad\reports\ReportingRequest();
+
+$repParams = '{
+    "startTime": "2016-01-01T00:00:00.000",
+    "endTime": "2017-10-01T00:00:00.000",
+    "selector": {
+    	"orderBy":[{"field":"campaignId","sortOrder":"DESCENDING"}]
+    },
+    "granularity":"MONTHLY"
+}';
+
+$rep->loadCertificates(__DIR__ . '/test.pem', __DIR__ . '/test.key');
+$cb = function ($params) use ($rep) {
+
+    var_dump("Callback for request mode");
+    if($rep->getRequestType() === 'R'){
+        $rep->setAllowRun(false);
+    }
+};
+
+$rep->addBeforeRequestCallback($cb, ['234']);
+
+$cond = new \searchad\selector\Conditions();
+//$cond->addCondition("campaignId", \searchad\selector\Conditions::OPERATOR_IN, ["9923026"]);
+//$cond->addCondition("modificationTime", \searchad\selector\Conditions::OPERATOR_LESS_THAN, ["2016-10-21T0:0:0.00"]);
+
+$res = $cond->getConditions();
+
+$sel = new \searchad\selector\Selector();
+
+$selData = $sel->orderBy("adGroupName")
+//    ->selectFields(["taps", "impressions"])
+    ->setLimit(3)
+    ->setOffset(0)
+    ->setConditions($res)
+    ->getSelector();
+
+$rep->setGranularity(\searchad\reports\ReportingRequest::GRANULARITY_DAILY)
+    ->setStartTime('2016-10-22')
+    ->setEndTime('2016-10-22')
+    ->setSelector($selData)
+    ->setReturnRowTotals(true)
+    ->queryReportsSearchTerm(9923026);
+
 $resp = new \searchad\ApiResponse();
+$resp->addCallback(function ($params) {
+    var_dump("response callback");
+    var_dump($params);
+}, []);
+$resp->loadResponse($rep->getRawResponse(), $rep->getCurlInfo());
 
 
-$api->loadCertificates(__DIR__ . '/kubrey-apple-ad.pem', __DIR__ . '/kubrey-apple-ad.key')
-    ->updateAdGroupInCampaign($id,$grid,json_encode(['status'=>'ENABLED']));
+var_dump(json_decode($rep->getRawResponse()), $rep->getRequestBody(true));
+//$api = new \searchad\campaign\CampaignRequest();
+
+//$api->addBeforeRequestCallback(function ($params) {
+//    var_dump("Callback");
+//    var_dump($params);
+//    $this->setAllowRun(false);
+//});
+//$resp = new \searchad\ApiResponse();
+
+
+//$api->loadCertificates(__DIR__ . '/kubrey-apple-ad.pem', __DIR__ . '/kubrey-apple-ad.key')
+//    ->updateAdGroupInCampaign($id, $grid, json_encode(['status' => 'ENABLED']));
 //    ->updateCampaign($id,json_encode(['status'=>'ENABLED']));
 
 //var_dump($api->getRawResponse());
-$resp->loadResponse($api->getRawResponse(),$api->getCurlInfo());
-var_dump($resp->httpCode(),$resp->getData(),$resp->getError());
+//$resp->loadResponse($api->getRawResponse(), $api->getCurlInfo());
+//var_dump($resp->httpCode(), $resp->getData(), $resp->getError());
 exit();
 
 //end of update
@@ -77,12 +137,11 @@ $rep->setGranularity(\searchad\reports\ReportingRequest::GRANULARITY_DAILY)
     ->queryReportsSearchTerm(9923026);
 
 $resp = new \searchad\ApiResponse();
-$resp->addCallback(function($params){
+$resp->addCallback(function ($params) {
     var_dump("response callback");
     var_dump($params);
-},[]);
-$resp->loadResponse($rep->getRawResponse(),$rep->getCurlInfo());
-
+}, []);
+$resp->loadResponse($rep->getRawResponse(), $rep->getCurlInfo());
 
 
 var_dump(json_decode($rep->getRawResponse()), $rep->getRequestBody(true));